spikes/spike2_deepgram/requirements.txt
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
